refactor: restructure backend project by removing unused files and updating Dockerfile for specific build path
