+++ backend/internal/models/book.go
