+++ backend/internal/models/streak_settings.go
