+++ backend/pkg/utils/middleware_helpers.go
@@ -1,4 +1,4 @@
-package middleware
+package utils
 
 import (
 	"crypto/rsa"
